Check to ensure database exists before updating card in migration
The `populate-card-database-id` migration was introduced to fix a bug
where the database id of native questions weren't be populated
correctly. If the the database associated to the native question is
removed, the question wouldn't be removed since there was no FK
relationship. This fix checks to ensure the database exists before
attempting to update the question with the database id. If it doesn't
exist, it cleans up the orphaned question.

Fixes #6520